patches for gensplash support from Chris Gianelloni <wolf31o2@gentoo.org>
git-svn-id: svn+ssh://svn.gentoo.org/var/svnroot/catalyst/trunk@466 d1e1f19c-881f-0410-ab34-b69fee027534

+++ livecd/runscript-support/livecdfs-update.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 # Copyright 1999-2004 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo/src/catalyst/livecd/runscript-support/Attic/livecdfs-update.sh,v 1.13 2004/10/15 02:41:03 zhen Exp $
+# $Header: /var/cvsroot/gentoo/src/catalyst/livecd/runscript-support/Attic/livecdfs-update.sh,v 1.14 2004/10/16 13:38:09 zhen Exp $
 
 /usr/sbin/env-update
 source /etc/profile
@@ -26,7 +26,7 @@ rc-update del consolefont
 rc-update add metalog default
 rc-update add modules default
 rc-update add pwgen default
-[ -e /etc/init.d/bootsplash ] && rc-update add bootsplash default
+[ -e /etc/init.d/splash ] && rc-update add splash default
 
 # switch the order of rcadd/ rcdel
 if [ -n "${clst_livecd_rcadd}" ] || [ -n "${clst_livecd_rcdel}" ]
@@ -51,11 +51,23 @@ fi
 rm -rf /etc/localtime
 cp /usr/share/zoneinfo/GMT /etc/localtime
 echo "livecd" > /etc/hostname
+echo "gentoo" > /etc/dnsdomainname
 
-sed -i -e '/\/dev\/[RBS]*/ s/^/#/' /etc/fstab
+# fstab tweaks
+#sed -i -e '/\/dev\/[RBS]*/ s/^/#/' /etc/fstab
 echo "tmpfs		/	tmpfs	defaults	0 0" >> /etc/fstab
 echo "tmpfs		/usr/lib/hotplug/firmware	tmpfs	defaults	0 0" >> /etc/fstab
 sed -i -e '/dev-state/ s/^/#/' /etc/devfsd.conf
+
+# tweak the livecd fstab so that users know not to edit it
+# http://bugs.gentoo.org/show_bug.cgi?id=60887
+mv /etc/fstab /etc/fstab.old
+echo "###############################################" >> /etc/fstab
+echo "## ATTENTION: THIS IS THE FSTAB ON THE LIVECD" >> /etc/fstab     
+echo "## PLEASE EDIT THE FSTAB IN /mnt/gentoo/etc" >> /etc/fstab     
+echo "###############################################" >> /etc/fstab
+cat /etc/fstab.old >> /etc/fstab
+rm /etc/fstab.old
 
 echo "alias cp='cp -i'" >> /etc/profile
 echo "alias ls='ls --color'" >> /etc/profile
@@ -66,8 +78,8 @@ echo "alias rm='rm -i'" >> /etc/profile
 # make sure we have the latest pci and hotplug ids
 if [ -d /usr/share/hwdata ]
 then
-	rm -f /usr/share/hwdata/pci.ids
-	rm -f /usr/share/hwdata/usb.ids
+	[ -f /usr/share/misc/pci.ids ] && rm -f /usr/share/hwdata/pci.ids
+	[ -f /usr/share/misc/usb.ids ] && rm -f /usr/share/hwdata/usb.ids
 	ln -s /usr/share/misc/pci.ids /usr/share/hwdata/pci.ids
 	ln -s /usr/share/misc/usb.ids /usr/share/hwdata/usb.ids
 fi
@@ -94,26 +106,28 @@ fi
 
 rm -f /etc/generic.motd.txt /etc/universal.motd.txt /etc/minimal.motd.txt /etc/gamecd.motd.txt
 
-# setup bootsplash (if called for)
-if [ -n "${clst_livecd_bootsplash}" ]
+# setup splash/bootsplash (if called for)
+if [ "${clst_livecd_splash_type}" == "bootsplash" -a -n "${clst_livecd_splash_theme}" ]
 then
-	if [ -d "/etc/bootsplash/${clst_livecd_bootsplash}" ]
+	if [ -d "/etc/bootsplash/${clst_livecd_splash_theme}" ]
 	then
 		sed -i 's/BOOTSPLASH_THEME=\"gentoo\"/BOOTSPLASH_THEME=\"${clst_livecd_bootsplash}\"/' /etc/conf.d/bootsplash
 		rm /etc/bootsplash/default
-		ln -s "/etc/bootsplash/${clst_livecd_bootsplash}" /etc/bootsplash/default
+		ln -s "/etc/bootsplash/${clst_livecd_splash_theme}" /etc/bootsplash/default
 	else
-		echo "Error, cannot setup bootsplash theme ${clst_livecd_bootsplash}"
+		echo "Error, cannot setup bootsplash theme ${clst_livecd_splash_theme}"
+		exit 1
+	fi
+
+elif [ "${clst_livecd_splash_type}" == "gensplash" -a -n "${clst_livecd_splash_theme}" ]
+then
+	if [ -d "/etc/splash/${clst_livecd_splash_theme}" ]
+	then
+		sed -i 's/# SPLASH_THEME="gentoo"/SPLASH_THEME=\"${clst_livecd_bootsplash}\"/' /etc/conf.d/splash
+		rm /etc/splash/default
+		ln -s "/etc/splash/${clst_livecd_splash_theme}" /etc/splash/default
+	else
+		echo "Error, cannot setup splash theme ${clst_livecd_splash_theme}"
 		exit 1
 	fi
 fi
-
-# tweak the livecd fstab so that users know not to edit it
-# http://bugs.gentoo.org/show_bug.cgi?id=60887
-mv /etc/fstab /etc/fstab.old
-echo "###############################################" >> /etc/fstab
-echo "## ATTENTION: THIS IS THE FSTAB ON THE LIVECD" >> /etc/fstab     
-echo "## PLEASE EDIT THE FSTAB IN /mnt/gentoo/etc" >> /etc/fstab     
-echo "###############################################" >> /etc/fstab
-cat /etc/fstab.old >> /etc/fstab
-rm /etc/fstab.old
